Prevent requested facing from being overridden by queued moves
If the move queue is not drained in real time, for example because
the game was suspended due to being in an out-of-focus tab,
the facing received in the `f` packet can get overwritten by
the one determined from the direction of an earlier queued move.
Delaying the SetFacing call until the queue is drained prevents
this from happening.

Player updates getting applied instantly, before the PlayerOther
is done moving, is an issue that applies to other packets as well
(e.g. sprite changes, so you have players sit-sliding to the spot
they've sat on), but facing is the only one the move queue can
override when not drained in real time.